refactor(public-api): redesign repository interfaces to accept pre-built FilterList
Repositories (daily-metrics, observations, traces, scores) now receive
pre-built FilterList objects instead of raw HTTP-param-shaped inputs.
Filter-building logic (convertApiProvidedFilterToClickhouseFilter,
deriveFilters, column mappings) moves to web-layer shim modules under
web/src/features/public-api/server/, keeping shared repos data-access-only.
