+++ packages/shared/src/server/repositories/observations.ts
@@ -16,8 +16,6 @@ import {
   StringFilter,
   FullObservations,
   orderByToClickhouseSql,
-  createPublicApiObservationsColumnMapping,
-  deriveFilters,
 } from "../queries";
 import { createFilterFromFilterState } from "../queries/clickhouse-sql/factory";
 import {
@@ -2119,31 +2117,19 @@ export const getCostByEvaluatorIds = async (
 
 // ─── Public-API observation query helpers ─────────────────────────────────────
 
-type PublicApiObservationQueryType = {
-  page: number;
-  limit: number;
+export const generateObservationsForPublicApi = async ({
+  projectId,
+  filter,
+  pagination,
+}: {
   projectId: string;
-  traceId?: string;
-  userId?: string;
-  name?: string;
-  type?: string;
-  parentObservationId?: string;
-  fromStartTime?: string;
-  toStartTime?: string;
-  version?: string;
-  advancedFilters?: FilterState;
-};
+  filter: FilterList;
+  pagination: { limit: number; page: number };
+}) => {
+  const appliedFilter = filter.apply();
+  const traceFilter = filter.find((f) => f.clickhouseTable === "traces");
 
-export const generateObservationsForPublicApi = async (
-  props: PublicApiObservationQueryType,
-) => {
-  const chFilter = generatePublicApiObservationFilter(props);
-  const appliedFilter = chFilter.apply();
-  const traceFilter = chFilter.find((f) => f.clickhouseTable === "traces");
-
-  const disableObservationsFinal = await shouldSkipObservationsFinal(
-    props.projectId,
-  );
+  const disableObservationsFinal = await shouldSkipObservationsFinal(projectId);
 
   const query = `
     with clickhouse_keys as (
@@ -2159,7 +2145,7 @@ export const generateObservationsForPublicApi = async (
         ${traceFilter ? `AND t.project_id = {projectId: String}` : ""}
         AND ${appliedFilter.query}
       ORDER BY start_time DESC
-        ${props.limit !== undefined && props.page !== undefined ? `LIMIT {limit: Int32} OFFSET {offset: Int32}` : ""}
+        LIMIT {limit: Int32} OFFSET {offset: Int32}
     )
     SELECT
       id,
@@ -2200,20 +2186,18 @@ export const generateObservationsForPublicApi = async (
 
   return measureAndReturn({
     operationName: "generateObservationsForPublicApi",
-    projectId: props.projectId,
+    projectId,
     input: {
       params: {
         ...appliedFilter.params,
-        projectId: props.projectId,
-        ...(props.limit !== undefined ? { limit: props.limit } : {}),
-        ...(props.page !== undefined
-          ? { offset: (props.page - 1) * props.limit }
-          : {}),
+        projectId,
+        limit: pagination.limit,
+        offset: (pagination.page - 1) * pagination.limit,
       },
       tags: {
         feature: "tracing",
         type: "observation",
-        projectId: props.projectId,
+        projectId,
         operation_name: "generateObservationsForPublicApi",
       },
     },
@@ -2229,12 +2213,15 @@ export const generateObservationsForPublicApi = async (
   });
 };
 
-export const getObservationsCountForPublicApi = async (
-  props: PublicApiObservationQueryType,
-) => {
-  const chFilter = generatePublicApiObservationFilter(props);
-  const filter = chFilter.apply();
-  const traceFilter = chFilter.find((f) => f.clickhouseTable === "traces");
+export const getObservationsCountForPublicApi = async ({
+  projectId,
+  filter,
+}: {
+  projectId: string;
+  filter: FilterList;
+}) => {
+  const appliedFilter = filter.apply();
+  const traceFilter = filter.find((f) => f.clickhouseTable === "traces");
 
   const query = `
     SELECT count() as count
@@ -2242,18 +2229,18 @@ export const getObservationsCountForPublicApi = async (
     ${traceFilter ? `LEFT JOIN __TRACE_TABLE__ t ON o.trace_id = t.id AND t.project_id = o.project_id` : ""}
     WHERE o.project_id = {projectId: String}
     ${traceFilter ? `AND t.project_id = {projectId: String}` : ""}
-    AND ${filter.query}
+    AND ${appliedFilter.query}
   `;
 
   return measureAndReturn({
     operationName: "getObservationsCountForPublicApi",
-    projectId: props.projectId,
+    projectId,
     input: {
-      params: { ...filter.params, projectId: props.projectId },
+      params: { ...appliedFilter.params, projectId },
       tags: {
         feature: "tracing",
         type: "observation",
-        projectId: props.projectId,
+        projectId,
         operation_name: "getObservationsCountForPublicApi",
       },
     },
@@ -2267,40 +2254,4 @@ export const getObservationsCountForPublicApi = async (
       return records.map((record) => Number(record.count)).shift();
     },
   });
-};
-
-const publicApiObservationsFilterParams =
-  createPublicApiObservationsColumnMapping(
-    "observations",
-    "o",
-    "parent_observation_id",
-  );
-
-const generatePublicApiObservationFilter = (
-  query: PublicApiObservationQueryType,
-) => {
-  const { advancedFilters, ...simpleFilterProps } = query;
-  const chFilter = deriveFilters(
-    simpleFilterProps,
-    publicApiObservationsFilterParams,
-    advancedFilters,
-    observationsTableUiColumnDefinitions.filter(
-      (c) => c.clickhouseTableName !== "scores",
-    ),
-    observationsTableCols,
-  );
-
-  const filteredChFilter = chFilter.filter(
-    (f) => f.clickhouseTable !== "scores",
-  );
-
-  filteredChFilter.push(
-    new StringFilter({
-      clickhouseTable: "observations",
-      field: "project_id",
-      operator: "=",
-      value: query.projectId,
-    }),
-  );
-  return filteredChFilter;
 };
