+++ packages/shared/src/server/repositories/scores.ts
@@ -25,12 +25,9 @@ import {
 import {
   FilterList,
   orderByToClickhouseSql,
-  StringFilter,
   StringOptionsFilter,
   DateTimeFilter,
   NumberFilter,
-  convertApiProvidedFilterToClickhouseFilter,
-  deriveFilters,
 } from "../queries";
 import { FilterCondition, FilterState, TimeFilter } from "../../types";
 import {
@@ -2410,211 +2407,25 @@ export type ScoreQueryType = {
   advancedFilters?: FilterState;
 };
 
-const secureScoreFilterOptions = [
-  {
-    id: "traceId",
-    clickhouseSelect: "trace_id",
-    clickhouseTable: "scores",
-    filterType: "StringFilter",
-    clickhousePrefix: "s",
-  },
-  {
-    id: "observationId",
-    clickhouseSelect: "observation_id",
-    clickhouseTable: "scores",
-    filterType: "StringOptionsFilter",
-    clickhousePrefix: "s",
-  },
-  {
-    id: "name",
-    clickhouseSelect: "name",
-    clickhouseTable: "scores",
-    filterType: "StringFilter",
-    clickhousePrefix: "s",
-  },
-  {
-    id: "source",
-    clickhouseSelect: "source",
-    clickhouseTable: "scores",
-    filterType: "StringFilter",
-    clickhousePrefix: "s",
-  },
-  {
-    id: "fromTimestamp",
-    clickhouseSelect: "timestamp",
-    operator: ">=" as const,
-    clickhouseTable: "scores",
-    filterType: "DateTimeFilter",
-    clickhousePrefix: "s",
-  },
-  {
-    id: "toTimestamp",
-    clickhouseSelect: "timestamp",
-    operator: "<" as const,
-    clickhouseTable: "scores",
-    filterType: "DateTimeFilter",
-    clickhousePrefix: "s",
-  },
-  {
-    id: "value",
-    clickhouseSelect: "value",
-    clickhouseTable: "scores",
-    filterType: "NumberFilter",
-    clickhousePrefix: "s",
-  },
-  {
-    id: "scoreIds",
-    clickhouseSelect: "id",
-    clickhouseTable: "scores",
-    filterType: "StringOptionsFilter",
-    clickhousePrefix: "s",
-  },
-  {
-    id: "configId",
-    clickhouseSelect: "config_id",
-    clickhouseTable: "scores",
-    filterType: "StringFilter",
-    clickhousePrefix: "s",
-  },
-  {
-    id: "sessionId",
-    clickhouseSelect: "session_id",
-    clickhouseTable: "scores",
-    filterType: "StringFilter",
-    clickhousePrefix: "s",
-  },
-  {
-    id: "datasetRunId",
-    clickhouseSelect: "dataset_run_id",
-    clickhouseTable: "scores",
-    filterType: "StringFilter",
-    clickhousePrefix: "s",
-  },
-  {
-    id: "queueId",
-    clickhouseSelect: "queue_id",
-    clickhouseTable: "scores",
-    filterType: "StringFilter",
-    clickhousePrefix: "s",
-  },
-  {
-    id: "environment",
-    clickhouseSelect: "environment",
-    clickhouseTable: "scores",
-    filterType: "StringOptionsFilter",
-    clickhousePrefix: "s",
-  },
-  {
-    id: "dataType",
-    clickhouseSelect: "data_type",
-    clickhouseTable: "scores",
-    filterType: "StringFilter",
-    clickhousePrefix: "s",
-  },
-];
-
-const secureTraceFilterOptions = [
-  {
-    id: "traceTags",
-    clickhouseSelect: "tags",
-    clickhouseTable: "traces",
-    filterType: "ArrayOptionsFilter",
-    clickhousePrefix: "t",
-  },
-  {
-    id: "userId",
-    clickhouseSelect: "user_id",
-    clickhouseTable: "traces",
-    filterType: "StringFilter",
-    clickhousePrefix: "t",
-  },
-];
-
-const determineTraceJoinRequirement = (
-  fields: string[] | null | undefined,
-  tracesFilterLength: number,
-) => {
-  const requestedFields = fields ?? ["score", "trace"];
-  const includeTrace = requestedFields.includes("trace");
-  const needsTraceJoin = includeTrace || tracesFilterLength > 0;
-  return { includeTrace, needsTraceJoin };
-};
-
-const generateScoreFilter = (
-  filter: ScoreQueryType,
-  scoreDataTypes?: readonly ScoreDataTypeType[],
-) => {
-  const scoresFilter = deriveFilters(
-    filter,
-    secureScoreFilterOptions,
-    filter.advancedFilters,
-    scoresTableUiColumnDefinitions,
-    scoresTableCols,
-  );
-  scoresFilter.push(
-    new StringFilter({
-      clickhouseTable: "scores",
-      field: "project_id",
-      operator: "=",
-      value: filter.projectId,
-    }),
-  );
-
-  if (scoreDataTypes) {
-    scoresFilter.push(
-      new StringOptionsFilter({
-        clickhouseTable: "scores",
-        field: "data_type",
-        operator: "any of",
-        values: [...scoreDataTypes],
-        tablePrefix: "s",
-      }),
-    );
-  }
-
-  const tracesFilter = convertApiProvidedFilterToClickhouseFilter(
-    filter,
-    secureTraceFilterOptions,
-  );
-
-  if (filter.environment && tracesFilter.length() > 0) {
-    const envValues = Array.isArray(filter.environment)
-      ? filter.environment
-      : [filter.environment];
-    tracesFilter.push(
-      new StringOptionsFilter({
-        clickhouseTable: "traces",
-        field: "environment",
-        operator: "any of",
-        values: envValues,
-        tablePrefix: "t",
-      }),
-    );
-  }
-
-  return { scoresFilter, tracesFilter };
-};
-
 export const _handleGenerateScoresForPublicApi = async ({
-  props,
+  projectId,
+  scoresFilter,
+  tracesFilter,
   scoreScope,
-  scoreDataTypes,
+  includeTrace,
+  needsTraceJoin,
+  pagination,
 }: {
-  props: ScoreQueryType;
+  projectId: string;
+  scoresFilter: FilterList;
+  tracesFilter: FilterList;
   scoreScope: "traces_only" | "all";
-  scoreDataTypes?: readonly ScoreDataTypeType[];
+  includeTrace: boolean;
+  needsTraceJoin: boolean;
+  pagination?: { limit: number; page: number };
 }) => {
-  const { scoresFilter, tracesFilter } = generateScoreFilter(
-    props,
-    scoreDataTypes,
-  );
   const appliedScoresFilter = scoresFilter.apply();
   const appliedTracesFilter = tracesFilter.apply();
-
-  const { includeTrace, needsTraceJoin } = determineTraceJoinRequirement(
-    props.fields,
-    tracesFilter.length(),
-  );
 
   const query = `
       SELECT
@@ -2670,26 +2481,28 @@ export const _handleGenerateScoresForPublicApi = async ({
           s.timestamp desc, s.event_ts desc
       LIMIT
           1 BY s.id, s.project_id
-      ${props.limit !== undefined && props.page !== undefined ? `LIMIT {limit: Int32} OFFSET {offset: Int32}` : ""}
+      ${pagination !== undefined ? `LIMIT {limit: Int32} OFFSET {offset: Int32}` : ""}
       `;
 
   return measureAndReturn({
     operationName: "_handleGenerateScoresForPublicApi",
-    projectId: props.projectId,
+    projectId,
     input: {
       params: {
         ...appliedScoresFilter.params,
         ...appliedTracesFilter.params,
-        projectId: props.projectId,
-        ...(props.limit !== undefined ? { limit: props.limit } : {}),
-        ...(props.page !== undefined
-          ? { offset: (props.page - 1) * props.limit }
+        projectId,
+        ...(pagination !== undefined
+          ? {
+              limit: pagination.limit,
+              offset: (pagination.page - 1) * pagination.limit,
+            }
           : {}),
       },
       tags: {
         feature: "scoring",
         type: "score",
-        projectId: props.projectId,
+        projectId,
         scoreScope,
         operation_name: "_handleGenerateScoresForPublicApi",
         includeTrace: includeTrace.toString(),
@@ -2730,25 +2543,22 @@ export const _handleGenerateScoresForPublicApi = async ({
 };
 
 export const _handleGetScoresCountForPublicApi = async ({
-  props,
+  projectId,
+  scoresFilter,
+  tracesFilter,
   scoreScope,
-  scoreDataTypes,
+  includeTrace,
+  needsTraceJoin,
 }: {
-  props: ScoreQueryType;
+  projectId: string;
+  scoresFilter: FilterList;
+  tracesFilter: FilterList;
   scoreScope: "traces_only" | "all";
-  scoreDataTypes?: readonly ScoreDataTypeType[];
+  includeTrace: boolean;
+  needsTraceJoin: boolean;
 }) => {
-  const { scoresFilter, tracesFilter } = generateScoreFilter(
-    props,
-    scoreDataTypes,
-  );
   const appliedScoresFilter = scoresFilter.apply();
   const appliedTracesFilter = tracesFilter.apply();
-
-  const { includeTrace, needsTraceJoin } = determineTraceJoinRequirement(
-    props.fields,
-    tracesFilter.length(),
-  );
 
   const query = `
       SELECT
@@ -2781,17 +2591,17 @@ export const _handleGetScoresCountForPublicApi = async ({
 
   return measureAndReturn({
     operationName: "_handleGetScoresCountForPublicApi",
-    projectId: props.projectId,
+    projectId,
     input: {
       params: {
         ...appliedScoresFilter.params,
         ...appliedTracesFilter.params,
-        projectId: props.projectId,
+        projectId,
       },
       tags: {
         feature: "scoring",
         type: "score",
-        projectId: props.projectId,
+        projectId,
         scoreScope,
         operation_name: "_handleGetScoresCountForPublicApi",
         includeTrace: includeTrace.toString(),
